workflow discovery import fixed in cli
